Added current RL78 17.0.1 sources (2496e989)

+++ libcxx/src/memory.cpp
@@ -192,7 +192,7 @@ align(size_t alignment, size_t size, void*& ptr, size_t& space)
     if (size <= space)
     {
         char* p1 = static_cast<char*>(ptr);
-        char* p2 = reinterpret_cast<char*>(reinterpret_cast<uintptr_t>(p1 + (alignment - 1)) & -alignment);
+        char* p2 = reinterpret_cast<char*>(reinterpret_cast<uintptr_t>(p1 + (alignment - 1)) & -static_cast<intptr_t>(alignment));
         size_t d = static_cast<size_t>(p2 - p1);
         if (d <= space - size)
         {
